Update metrics dependency to 0.17
This change updates the metrics dependency to 0.17, which requires
several code changes around gauge value type and a mock recorder
instance.

+++ libsawtooth/src/journal/block_scheduler.rs
@@ -236,11 +236,11 @@ impl<B: BlockStatusStore> BlockSchedulerState<B> {
     fn update_gauges(&self) {
         gauge!(
             "block_validator.BlockScheduler.blocks_processing",
-            self.processing.len() as i64
+            self.processing.len() as f64
         );
         gauge!(
             "block_validator.BlockScheduler.blocks_pending",
-            self.pending.len() as i64
+            self.pending.len() as f64
         );
     }
 }
